pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vnstock_ai_agent"
version = "0.1.0"
description = "Vietnamese Stock AI Advisor Agent with HippoRAG memory"
requires-python = ">=3.10"

dependencies = [
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.29",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "groq>=0.11.0",
    "openai>=1.0.0",
    "tqdm>=4.66.0",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
